Move orig steps to the start of the temp file
Unlike FRAT, the LRAT format implicitly starts with all orig steps added
to the state. To address this mismatch, we first push all orig steps to
the beginning of the temp file, then read them in as a block during the
second phase. Additionally, orig steps can be merged if they happen to
be permutations of each other, so we keep a count of how many copies
of a given orig step are in the state, deleting only once the last copy is
removed.