add MidVec data structure
This was implicitly being used in drat-trim already, but here I give MidVec
a proper API. A MidVec is like a Box<[T]>, in that it is a
pointer-length pair, but the pointer actually points to the center of
the allocation and the length is the "radius" of the allocation; the actual
number of allocated elements is `2*n+1`. This is useful for operations
indexed on literals, because we can just index by the literal number
with a single addition.

Other changes:
* Instead of using thread local storage, we just pass in a `Propagate` struct
  containing work variables.
* Instead of returning a `Hint` object, which requires some unnecessary
 allocation, we store the data inside a vector in `Propagate` that can
 be reused. The actual return value of `propagate` is now just a bool
 signalling success.
* The variable assignment is now a `MidVec`, where the bool vector
  indicates whether a literal is assigned to true vs unassigned, with
  assignment to false being indicated by assigning the complement
  literal to true. The reasons are stored along with the variable
  assignment now, together with a stack of assigned literals so that we
  can quickly clear the vector without having to traverse the whole
  thing if the number of literals is large.