more perf tricks from drat-trim
The main change here is to not clear the variable assignment between
calls to `propagate`. The propagation of top level units can be mostly
retained across calls, and this should give a nice speedup.

I tried to remove the `units` hashmap, in favor of the units stored in
the `tru_stack`, but I was ultimately unsuccessful; it is very difficult
to do something both cheap and correct in `unassign` because we have to
recheck all the units downstream of the removed one to make sure that
they aren't invalidated, and even if they are we have to make sure not
to drop them if they can be derived by alternate means (for example if
there are two copies of a unit literal in the context).

For now, we use the `units_processed` flag to indicate that units may
need to be refreshed from the `units` hashmap.

Other changes:
* The `tru_lits` map now stores marks in the same place as the assignment
 itself; marks are set in `propagate` and `propagate_hint` and cleared
  right afterward in `finalize_hint`.
* `first_hyp` tracks the index of the first hypothesis, so that we can
  clear hypotheses while retaining global settings.
  This is `S->forced` from drat-trim.
* `first_unprocessed` is `S->processed` in drat-trim; it tracks the
  initial segment of the true stack such that all consequences of the
  literals before the bar are in the true stack. Note that
   `propagate_hint` does not update this value, because it does not
   calculate all consequences, only the ones specified in the hint.
* The variable assignment can also end up in "unsat mode", where
  further propagations are forbidden. In this state, the last item in the
  true stack is a literal that is both true and false (and no other
  literals have this property).
  When propagate sees that the assignment is in unsat mode, it just
  re-marks the conflict literals and returns unsat, meaning that we will
  use this conflict to prove anything we see.
* `Watches` is an encapsulation of the watch list, just to make it easier
  to use. The ability to lazily initialize the watch list has been
  dropped, because it is needed in `run_rat_step` anyway,
  so the only way you can avoid needing a watch list is if all steps are
  annotated and there are no RAT steps.
* The `self_test` function is just for debugging; it will check that the
  context is in a vaild state (and is useful besides as a way to document
  the invariants).
* `propagate` has been split into `propagate_core`, which just does
  propagation of the current variable assignment, and `propagate` which
  first propagates globals and then the hypotheses.